// src/core/models/CoinModel.ts
export interface Coin {
    id: string;
    csupply: string;
    market_cap_usd: string;
    msupply: string;
    name: string;
    nameid: string;
    percent_change_1h: string;
    percent_change_7d: string;
    percent_change_24h: string;
    price_btc: string;
    price_usd: string;
    rank: number;
    symbol: string;
    tsupply: string;
    volume24: number;
    volume24a: number;
}